+++ lending/loan_management/doctype/loan_application/test_loan_application.py
@@ -7,7 +7,7 @@ import frappe
 
 from erpnext.setup.doctype.employee.test_employee import make_employee
 
-from lending.loan_management.doctype.loan.test_utils import (
+from lending.tests.test_utils import (
 	create_loan_accounts,
 	create_loan_product,
 	set_loan_settings_in_company,
